+++ Microsoft.ReportViewer.WinForms/Microsoft.Reporting.WinForms.Internal.Soap.ReportingServices2005.Execution/ReportExecutionServiceSoapClientExtension.cs
@@ -23,7 +23,7 @@ namespace Microsoft.Reporting.WinForms.Internal.Soap.ReportingServices2005.Execu
 				ClientCredentials.Windows.ClientCredential.Password = cred.Password;
 				ClientCredentials.Windows.AllowedImpersonationLevel = System.Security.Principal.TokenImpersonationLevel.Delegation;
 				var binding = (BasicHttpBinding)Endpoint.Binding;
-				binding.Security.Mode = BasicHttpSecurityMode.TransportCredentialOnly;
+				binding.Security.Mode = Endpoint.Address.Uri.Scheme == "https" ? BasicHttpSecurityMode.Transport : BasicHttpSecurityMode.TransportCredentialOnly;
 				binding.Security.Transport.ClientCredentialType = HttpClientCredentialType.Ntlm;
 			}
 		}