+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # upcoming version
  * Added missing System.Text.RegularExpressions.dll assembly reference
  * https support for server-side reports
+ * Fixed chart HTML4/5 rendering
 
 # 15.1.13
  * Fixed BinaryFormatter/ResX issues with PDF export on .NET 5